payment-worker/requirements.txt
pika
psycopg2-binary
pymongo